neuro_market/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
websockets
yfinance
numpy
pandas
scikit-learn
deepface
opencv-python
psycopg2-binary
requests
python-dotenv
email-validator
tf-keras
matplotlib
reportlab
pymysql
